use dialogflow namespace for types
